don't abort immediately when a NaN/inf is encountered; emit warnings instead
Current behavior of SRProxy is to abort() with an error whenever an inf/NaN is encountered.
While this aids in finding infs/NaNs, it makes dealing with files that have these problematic values baked in difficult in the period in between discovery and reproduction of the files.
Switch behavior to tabulating infs/NaNs and emitting a summary at the end of the run.
The old behavior is still accessible by setting environment var SRPROXY_ABORT_ON_INFNAN=1.